refactor(polars): Move compat functions out of `polars.expr`
Aiming to implement enough of `PolarsExpr` to get all/most
of these tests working in `tests.plan.compliant_test.py`:
- `test_select`
- `test_with_columns`
- `test_with_columns_all_aggregates`
- `test_first_last_expr_with_columns`

Need more room for that
