+++ src/narwhals/_plan/polars/expr.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import builtins
 from typing import TYPE_CHECKING, Any, ClassVar, TypeVar
 
 import polars as pl
@@ -8,13 +7,12 @@ import polars as pl
 from narwhals._plan.common import todo
 from narwhals._plan.compliant import CompliantExpr, typing as ct
 from narwhals._plan.compliant.accessors import ExprStructNamespace
-from narwhals._plan.polars import compat
+from narwhals._plan.polars import functions as fn
 from narwhals._plan.polars.classes import PolarsClasses
 from narwhals._plan.polars.namespace import dtype_to_native, dtype_to_native_fast
 from narwhals._utils import Version
 
 if TYPE_CHECKING:
-    from collections.abc import Sequence
     from typing import TypeAlias
 
     from typing_extensions import Self
@@ -26,86 +24,11 @@ if TYPE_CHECKING:
     from narwhals._plan.polars.dataframe import PolarsDataFrame as DataFrame  # noqa: F401
     from narwhals.typing import IntoDType, PythonLiteral
 
-__all__ = ("PolarsExpr", "len", "linear_space", "lit", "over", "row_index")
+__all__ = ("PolarsExpr",)
 
 Incomplete: TypeAlias = Any
 
 PolarsFrame: TypeAlias = "ct.Frame[pl.DataFrame, pl.Series, pl.LazyFrame]"
-
-if compat.OVER_RESPECTS_NULLS_LAST:
-    # NOTE: Allows all features, so no need to branch in any calls
-    def over(
-        self: pl.Expr,
-        *partition_by: pl.Expr | str,
-        order_by: Sequence[str] | None = None,
-        descending: bool = False,
-        nulls_last: bool = False,
-    ) -> pl.Expr:
-        return self.over(
-            *partition_by, order_by=order_by, descending=descending, nulls_last=nulls_last
-        )
-else:
-
-    def over(
-        self: pl.Expr,
-        *partition_by: pl.Expr | str,
-        order_by: Sequence[str] | None = None,
-        descending: bool = False,
-        nulls_last: bool = False,
-    ) -> pl.Expr:
-        if nulls_last:
-            raise compat.over_error("nulls_last")
-        options: dict[str, Any] = {}
-        if order_by:
-            if not compat.OVER_SUPPORTS_ORDER_BY:
-                raise compat.over_error("order_by_any")
-            options["order_by"] = order_by
-            if descending:
-                if not compat.OVER_SUPPORTS_DESCENDING:
-                    raise compat.over_error("descending")
-                options["descending"] = descending
-            if not partition_by and not compat.OVER_WITHOUT_PARTITION_BY:
-                partition_by = (lit(1),)
-        return self.over(*partition_by, **options)
-
-
-if compat.HAS_LINEAR_SPACE or TYPE_CHECKING:
-    # NOTE: Has some pretty intricate `@overload`s that can be preserved this way
-    linear_space = pl.linear_space
-else:
-
-    def linear_space(*_: Any, **__: Any) -> Any:
-        raise compat.too_old("linear_space", "1.21.0")
-
-
-if compat.HAS_LEN or TYPE_CHECKING:
-    len = pl.len
-else:
-
-    def len() -> pl.Expr:
-        return pl.count().alias("len")
-
-
-def row_index(
-    name: str = "index", order_by: Sequence[str] = (), *, nulls_last: bool = False
-) -> pl.Expr:
-    int_range = pl.int_range(len()).alias(name)
-    if not order_by:
-        return int_range
-    if compat.OVER_RESPECTS_NULLS_LAST:
-        return int_range.over(order_by=order_by, nulls_last=nulls_last)
-    # NOTE: `nulls_last` isn't the missing feature,
-    # but the behavior is more predictable following that change
-    by = pl.col(order_by) if builtins.len(order_by) == 1 else pl.struct(order_by)
-    return int_range.sort_by(by.arg_sort(nulls_last=nulls_last))
-
-
-if compat.LIT_ACCEPTS_DICT or TYPE_CHECKING:
-    lit = pl.lit
-else:
-
-    def lit(value: Any, dtype: pl.DataType | type[pl.DataType] | None = None) -> pl.Expr:
-        return pl.struct(**value) if isinstance(value, dict) else pl.lit(value, dtype)
 
 
 ExprT_co = TypeVar("ExprT_co", bound="PolarsExpr", covariant=True)
@@ -137,7 +60,7 @@ class PolarsExpr(CompliantExpr["DataFrame", pl.Expr, pl.Expr]):
     ) -> Self:
         unknown = cls.version.dtypes.Unknown
         dtype_pl = None if dtype == unknown else dtype_to_native(dtype, cls.version)
-        return cls.from_native(lit(value, dtype_pl), name)
+        return cls.from_native(fn.lit(value, dtype_pl), name)
 
     @property
     def native(self) -> pl.Expr:
@@ -167,11 +90,11 @@ class PolarsExpr(CompliantExpr["DataFrame", pl.Expr, pl.Expr]):
     def lit_series(
         cls, node: ir.LitSeries[pl.Series], _: Incomplete, name: str, /
     ) -> Self:
-        return cls.from_native(lit(node.native), name)
+        return cls.from_native(fn.lit(node.native), name)
 
     @classmethod
     def len_star(cls, _: ir.Len, __: Incomplete, name: str, /) -> Self:
-        return cls.from_native(len(), name)
+        return cls.from_native(fn.len(), name)
 
     abs = todo()
     all = todo()
