+++ src/narwhals/_plan/polars/series.py
@@ -8,9 +8,8 @@ import polars as pl
 from narwhals._plan._version import into_version
 from narwhals._plan.compliant import CompliantSeries, typing as ct
 from narwhals._plan.compliant.accessors import SeriesStructNamespace as StructNamespace
-from narwhals._plan.polars import compat
+from narwhals._plan.polars import compat, functions as fn
 from narwhals._plan.polars.classes import PolarsClasses
-from narwhals._plan.polars.expr import linear_space, lit
 from narwhals._plan.polars.namespace import (
     dtype_from_native,
     dtype_to_native,
@@ -243,7 +242,7 @@ class PolarsSeries(CompliantSeries[pl.Series]):
         return None if self.is_empty() else self.native.item(-1)
 
     def fill_nan(self, value: float | Self | None) -> Self:
-        fill_value = lit(value.native) if isinstance(value, PolarsSeries) else value
+        fill_value = fn.lit(value.native) if isinstance(value, PolarsSeries) else value
         return self._with_native(self.native.fill_nan(fill_value))
 
     def fill_null(self, value: NonNestedLiteral | Self) -> Self:
@@ -458,7 +457,7 @@ class PolarsSeries(CompliantSeries[pl.Series]):
         closed: ClosedInterval = "both",
         name: str = "literal",
     ) -> Self:
-        native = linear_space(start, end, num_samples, closed=closed, eager=True)
+        native = fn.linear_space(start, end, num_samples, closed=closed, eager=True)
         return cls.from_native(native, name)
 
     @property
